+++ dtx/sagas/registry_test.go
@@ -14,16 +14,14 @@ func TestRegistry_Recover(t *testing.T) {
 	store := NewInProcessStore()
 	store.transactions["test"] = []Log{{
 		ID:            "0",
-		correlationID: "2",
+		CorrelationID: "2",
 		StartedAt:     time.Now(),
 		LogType:       Session,
-		StepNumber:    0,
 	}, {
 		ID:            "1",
-		correlationID: "2",
+		CorrelationID: "2",
 		StartedAt:     time.Now(),
 		FinishedAt:    time.Time{},
-		StepNumber:    1,
 		LogType:       Do,
 		StepError:     nil,
 	}}
@@ -35,15 +33,14 @@ func TestRegistry_RecoverWithTimeout(t *testing.T) {
 	store := NewInProcessStore()
 	store.transactions["test"] = []Log{{
 		ID:            "0",
-		correlationID: "2",
+		CorrelationID: "2",
 		StartedAt:     time.Now(),
 		LogType:       Session,
 	}, {
 		ID:            "1",
-		correlationID: "2",
+		CorrelationID: "2",
 		StartedAt:     time.Now(),
 		FinishedAt:    time.Time{},
-		StepNumber:    0,
 		LogType:       Do,
 		StepError:     nil,
 		StepName:      "foo",
@@ -67,15 +64,14 @@ func TestRegistry_RecoverSerialized(t *testing.T) {
 	store := NewInProcessStore()
 	store.transactions["test"] = []Log{{
 		ID:            "0",
-		correlationID: "2",
+		CorrelationID: "2",
 		StartedAt:     time.Now().Add(-time.Hour),
 		LogType:       Session,
 	}, {
 		ID:            "1",
-		correlationID: "2",
+		CorrelationID: "2",
 		StartedAt:     time.Now().Add(-time.Hour),
 		FinishedAt:    time.Time{},
-		StepNumber:    0,
 		LogType:       Do,
 		StepError:     nil,
 		StepName:      "foo",