+++ ots3/uploader_test.go
@@ -5,6 +5,7 @@ package ots3
 import (
 	"context"
 	"net/http"
+	"os"
 	"testing"
 
 	"github.com/DoNewsCode/core/key"
@@ -14,23 +15,6 @@ import (
 	"github.com/opentracing/opentracing-go/mocktracer"
 	"github.com/stretchr/testify/assert"
 )
-
-func setupManager() *Manager {
-	return setupManagerWithTracer(nil)
-}
-
-func setupManagerWithTracer(tracer opentracing.Tracer) *Manager {
-	m := NewManager(
-		"Q3AM3UQ867SPQQA43P2F",
-		"zuf+tfteSlswRu7BJ86wekitnifILbZam1KYY3TG",
-		"https://play.minio.io:9000",
-		"asia",
-		"mybucket",
-		WithTracer(tracer),
-	)
-	_ = m.CreateBucket(context.Background(), "mybucket")
-	return m
-}
 
 func TestNewManager(t *testing.T) {
 	t.Parallel()
@@ -50,9 +34,15 @@ func TestNewManager(t *testing.T) {
 	))
 }
 
+func TestMain(m *testing.M) {
+	manager := NewManager("Q3AM3UQ867SPQQA43P2F", "zuf+tfteSlswRu7BJ86wekitnifILbZam1KYY3TG", "https://play.minio.io:9000", "asia", "mybucket")
+	_ = manager.CreateBucket(context.Background(), "foo")
+	os.Exit(m.Run())
+}
+
 func TestManager_CreateBucket(t *testing.T) {
 	t.Parallel()
-	m := NewManager("Q3AM3UQ867SPQQA43P2F", "zuf+tfteSlswRu7BJ86wekitnifILbZam1KYY3TG", "https://play.minio.io:9000", "asia", "mybucket")
+	m := setupManager()
 	err := m.CreateBucket(context.Background(), "foo")
 	if err != nil {
 		if aerr, ok := err.(awserr.Error); ok {
@@ -74,8 +64,25 @@ func TestManager_CreateBucket(t *testing.T) {
 func TestManager_UploadFromUrl(t *testing.T) {
 	tracer := mocktracer.New()
 	m := setupManagerWithTracer(tracer)
+	_ = m.CreateBucket(context.Background(), "mybucket")
 	newURL, err := m.UploadFromUrl(context.Background(), "https://www.donews.com/static/v2/images/full-logo.png")
 	assert.NoError(t, err)
 	assert.NotEmpty(t, newURL)
 	assert.Len(t, tracer.FinishedSpans(), 2)
+}
+
+func setupManager() *Manager {
+	return setupManagerWithTracer(nil)
+}
+
+func setupManagerWithTracer(tracer opentracing.Tracer) *Manager {
+	m := NewManager(
+		"Q3AM3UQ867SPQQA43P2F",
+		"zuf+tfteSlswRu7BJ86wekitnifILbZam1KYY3TG",
+		"https://play.minio.io:9000",
+		"asia",
+		"mybucket",
+		WithTracer(tracer),
+	)
+	return m
 }